Addd benchmark througput script and runner
