+++ neuralmagic/benchmarks/configs/benchmark_serving.json
@@ -1,16 +1,20 @@
 {
-	"description" : "Benchmarking system test",
+	"configs" : [
+		{
+			"description" : "Benchmark vllm serving",
+		
+			"dataset_download_cmds" : ["wget https://huggingface.co/datasets/anon8231489123/ShareGPT_Vicuna_unfiltered/resolve/main/ShareGPT_V3_unfiltered_cleaned_split.json"],
+		
+			"models" : ["facebook/opt-125m"],
+		
+			"script_name" : "benchmark_serving.py",
 
-	"dataset_download_cmds" : ["wget https://huggingface.co/datasets/anon8231489123/ShareGPT_Vicuna_unfiltered/resolve/main/ShareGPT_V3_unfiltered_cleaned_split.json"],
-
-	"models" : ["facebook/opt-125m"],
-
-	"script_name" : "benchmark_serving.py",
-	
-	"script_args" : {
-		"num-prompts" : [20, 25],
-		"request-rate" : [5, 10],
-		"best-of" : [1],
-		"dataset" : ["ShareGPT_V3_unfiltered_cleaned_split.json"] 
-	}
+			"script_args" : {
+				"num-prompts" : [20, 25],
+				"request-rate" : [5, 10],
+				"best-of" : [1],
+				"dataset" : ["ShareGPT_V3_unfiltered_cleaned_split.json"] 
+			}
+		}
+	]
 }
